+++ characters/watermelon.py
@@ -153,7 +153,7 @@ class Watermelon(Sprite):
     def draw_player(self, screen):
         rotated_image = pygame.transform.rotate(self.image, self.angle - 180)
         w, h = rotated_image.get_size()
-        camera.blit(rotated_image, (int(self.x - w / 2), int(self.y - h / 2)))
+        camera.blit(rotated_image, (self.x - w / 2, self.y - h / 2))
 
     def draw_flame(self, screen):
         image_height = int(abs(self.velocity)) + 10
@@ -162,8 +162,8 @@ class Watermelon(Sprite):
         w, h = rotated_image.get_size()
 
         distance_from_center = (20 + int(abs(self.velocity) / 2))
-        camera.blit(rotated_image, (int(self.x - distance_from_center * self.orientation_vector_x - w / 2),
-            int(self.y - distance_from_center * self.orientation_vector_y - h / 2)))
+        camera.blit(rotated_image, (self.x - distance_from_center * self.orientation_vector_x - w / 2,
+            self.y - distance_from_center * self.orientation_vector_y - h / 2))
 
     def draw_hud(self, screen):
         # health bar
